requirements.txt
fastapi
uvicorn
torch
markdown
scikit-learn
beautifulsoup4
python-dotenv
pytest
requests
langchain
supabase
unstructured
langchain-huggingface
sentence-transformers==2.7.0
transformers==4.41.2
